+++ php/modules/Photo.php
@@ -41,7 +41,7 @@ class Photo extends Module {
 	public function add($files, $albumID, $description = '', $tags = '') {
 
 		# Check dependencies
-		$this->dependencies(isset($this->database));
+		self::dependencies(isset($this->database));
 
 		# Check permissions
 		if (hasPermissions(LYCHEE_UPLOADS_BIG)===false||hasPermissions(LYCHEE_UPLOADS_THUMB)===false) {
@@ -65,6 +65,13 @@ class Photo extends Module {
 				# f for starred (fav)
 				$star		= 1;
 				$public		= 0;
+				$albumID	= 0;
+				break;
+
+			case 'r':
+				# r for recent
+				$public		= 0;
+				$star		= 0;
 				$albumID	= 0;
 				break;
 
@@ -130,7 +137,7 @@ class Photo extends Module {
 			}
 
 			# Save to DB
-			$query = "INSERT INTO lychee_photos (id, title, url, description, tags, type, width, height, size, iso, aperture, make, model, shutter, focal, takestamp, thumbUrl, album, public, star)
+			$query = "INSERT INTO lychee_photos (id, title, url, description, tags, type, width, height, size, iso, aperture, make, model, shutter, focal, takestamp, thumbUrl, album, public, star, checksum)
 				VALUES (
 					'" . $id . "',
 					'" . $info['title'] . "',
@@ -151,7 +158,8 @@ class Photo extends Module {
 					'" . md5($id) . ".jpeg',
 					'" . $albumID . "',
 					'" . $public . "',
-					'" . $star . "');";
+					'" . $star . "',
+					'" . md5_file($path) . "');";
 			$result = $this->database->query($query);
 
 			if (!$result) {
@@ -171,7 +179,7 @@ class Photo extends Module {
 	private function createThumb($url, $filename, $width = 200, $height = 200) {
 
 		# Check dependencies
-		$this->dependencies(isset($this->database, $this->settings, $url, $filename));
+		self::dependencies(isset($this->database, $this->settings, $url, $filename));
 
 		# Call plugins
 		$this->plugins(__METHOD__, 0, func_get_args());
@@ -255,7 +263,7 @@ class Photo extends Module {
 	private function adjustFile($path, $info) {
 
 		# Check dependencies
-		$this->dependencies(isset($path, $info));
+		self::dependencies(isset($path, $info));
 
 		# Call plugins
 		$this->plugins(__METHOD__, 0, func_get_args());
@@ -369,7 +377,7 @@ class Photo extends Module {
 	public function get($albumID) {
 
 		# Check dependencies
-		$this->dependencies(isset($this->database, $this->photoIDs));
+		self::dependencies(isset($this->database, $this->photoIDs));
 
 		# Call plugins
 		$this->plugins(__METHOD__, 0, func_get_args());
@@ -381,6 +389,9 @@ class Photo extends Module {
 		# Parse photo
 		$photo['sysdate'] = date('d M. Y', substr($photo['id'], 0, -4));
 		if (strlen($photo['takestamp'])>1) $photo['takedate'] = date('d M. Y', $photo['takestamp']);
+
+		# Parse url
+		$photo['url'] = LYCHEE_URL_UPLOADS_BIG . $photo['url'];
 
 		if ($albumID!='false') {
 
@@ -410,7 +421,7 @@ class Photo extends Module {
 	private function getInfo($url) {
 
 		# Check dependencies
-		$this->dependencies(isset($this->database, $url));
+		self::dependencies(isset($this->database, $url));
 
 		# Call plugins
 		$this->plugins(__METHOD__, 0, func_get_args());
@@ -504,7 +515,7 @@ class Photo extends Module {
 	public function getArchive() {
 
 		# Check dependencies
-		$this->dependencies(isset($this->database, $this->photoIDs));
+		self::dependencies(isset($this->database, $this->photoIDs));
 
 		# Call plugins
 		$this->plugins(__METHOD__, 0, func_get_args());
@@ -541,7 +552,7 @@ class Photo extends Module {
 	public function setTitle($title) {
 
 		# Check dependencies
-		$this->dependencies(isset($this->database, $this->photoIDs));
+		self::dependencies(isset($this->database, $this->photoIDs));
 
 		# Call plugins
 		$this->plugins(__METHOD__, 0, func_get_args());
@@ -566,7 +577,7 @@ class Photo extends Module {
 	public function setDescription($description) {
 
 		# Check dependencies
-		$this->dependencies(isset($this->database, $this->photoIDs));
+		self::dependencies(isset($this->database, $this->photoIDs));
 
 		# Call plugins
 		$this->plugins(__METHOD__, 0, func_get_args());
@@ -592,7 +603,7 @@ class Photo extends Module {
 	public function setStar() {
 
 		# Check dependencies
-		$this->dependencies(isset($this->database, $this->photoIDs));
+		self::dependencies(isset($this->database, $this->photoIDs));
 
 		# Call plugins
 		$this->plugins(__METHOD__, 0, func_get_args());
@@ -629,7 +640,7 @@ class Photo extends Module {
 	public function getPublic($password) {
 
 		# Check dependencies
-		$this->dependencies(isset($this->database, $this->photoIDs));
+		self::dependencies(isset($this->database, $this->photoIDs));
 
 		# Call plugins
 		$this->plugins(__METHOD__, 0, func_get_args());
@@ -657,7 +668,7 @@ class Photo extends Module {
 	public function setPublic() {
 
 		# Check dependencies
-		$this->dependencies(isset($this->database, $this->photoIDs));
+		self::dependencies(isset($this->database, $this->photoIDs));
 
 		# Call plugins
 		$this->plugins(__METHOD__, 0, func_get_args());
@@ -686,7 +697,7 @@ class Photo extends Module {
 	function setAlbum($albumID) {
 
 		# Check dependencies
-		$this->dependencies(isset($this->database, $this->photoIDs));
+		self::dependencies(isset($this->database, $this->photoIDs));
 
 		# Call plugins
 		$this->plugins(__METHOD__, 0, func_get_args());
@@ -708,7 +719,7 @@ class Photo extends Module {
 	public function setTags($tags) {
 
 		# Check dependencies
-		$this->dependencies(isset($this->database, $this->photoIDs));
+		self::dependencies(isset($this->database, $this->photoIDs));
 
 		# Call plugins
 		$this->plugins(__METHOD__, 0, func_get_args());
@@ -738,7 +749,7 @@ class Photo extends Module {
 	public function delete() {
 
 		# Check dependencies
-		$this->dependencies(isset($this->database, $this->photoIDs));
+		self::dependencies(isset($this->database, $this->photoIDs));
 
 		# Call plugins
 		$this->plugins(__METHOD__, 0, func_get_args());